Fixed some bugs..

+++ App/app/src/main/java/com/cmov/bomberman/model/agent/Agent.java
@@ -78,7 +78,7 @@ public abstract class Agent {
 	 * @param state the game state
 	 * @param dt the time since the last update
 	 */
-	abstract public void play(final State state, final long dt);
+	abstract public void play(final State state, final float dt);
 
 	/**
 	 * Returns if this agent is already destroyed (should be removed from the game)
